% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annoRBP}
\alias{annoRBP}
\title{RNA binding protein function annotations, localization patterns, and predicted RNA binding domains.}
\format{
data.frame
}
\usage{
annoRBP
}
\description{
RNA binding protein function annotations, localization patterns, and predicted RNA binding domains.
}
\note{
Table S2 of PMID 32252787

there is a mix of types even though data are 0/1 ... because of excel NA token.
}
\keyword{datasets}
