% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoDE}
\alias{demoDE}
\title{a DESeq2 result for ZNF317-BGRSDLV02-RS293_DESeq2}
\format{
data.frame
}
\usage{
demoDE
}
\description{
a DESeq2 result for ZNF317-BGRSDLV02-RS293_DESeq2
}
\keyword{datasets}
