% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanceDE.R
\name{enhanceDE}
\alias{enhanceDE}
\title{convert a data.frame with DESeq2 results from an ENCORE experiment to an annotated GRanges}
\usage{
enhanceDE(deframe, knockdown)
}
\arguments{
\item{deframe}{a data.frame instance with DESeq2 results}

\item{knockdown}{character(1) name of gene knocked down}
}
\description{
convert a data.frame with DESeq2 results from an ENCORE experiment to an annotated GRanges
}
\examples{
data(demoDE)
dat = enhanceDE(demoDE, "ZNF317")
dat[order(dat$padj)]
}
