% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode.v40.fullgene}
\alias{gencode.v40.fullgene}
\title{an excerpt from GENCODE v40 catalog limited to type "gene"}
\format{
GRanges
}
\usage{
gencode.v40.fullgene
}
\description{
an excerpt from GENCODE v40 catalog limited to type "gene"
}
\note{
GRCh38 coordinates; a token "NT" in `rbp_cands` is not matched
}
\keyword{datasets}
