% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rMATS.R
\name{get_rMATS_tgz}
\alias{get_rMATS_tgz}
\title{retrieve path to an rMATS gzipped tar archive after retrieving from DRS Hub}
\usage{
get_rMATS_tgz(target = "CTCF", outputBasename = tempfile())
}
\arguments{
\item{target}{character(1) a symbol for a gene coding for an RNA binding protein, defaults to "CTCF",
which likely needs to be combined with a disambiguating string regexp"}

\item{outputBasename}{character(1) string to which .tgz will be appended after retrieval}
}
\value{
characer(1) path to .tgz result
}
\description{
retrieve path to an rMATS gzipped tar archive after retrieving from DRS Hub
}
\examples{
ex = "ZNF800-BGRSLV02-RS293"
path = get_rMATS_tgz(ex)
path
tf = tempfile()
untar(path, exdir=tf)
lkd = dir(tf, recursive=TRUE, full=TRUE)
a3SS = grep("A3SS", lkd, value=TRUE)
z = data.table::fread(a3SS)
length(unique(z$geneSymbol[z$PValue < .001]))
}
