% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rbp_cands_gr}
\alias{rbp_cands_gr}
\title{an excerpt from GENCODE v40 catalog limited to "gene" entries for the `rbp_cands`, as GRanges instance}
\format{
GRanges
}
\usage{
rbp_cands_gr
}
\description{
an excerpt from GENCODE v40 catalog limited to "gene" entries for the `rbp_cands`, as GRanges instance
}
\note{
GRCh38 coordinates; a token "NT" in `rbp_cands` is not matched
}
\keyword{datasets}
