% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rbp_cands_txex}
\alias{rbp_cands_txex}
\title{an excerpt from GENCODE v40 catalog with `gene_name` matching `rbp_cands`; all types (transcripts, exons, ...), as GRanges instance}
\format{
GRanges
}
\usage{
rbp_cands_txex
}
\description{
an excerpt from GENCODE v40 catalog with `gene_name` matching `rbp_cands`; all types (transcripts, exons, ...), as GRanges instance
}
\note{
GRCh38 coordinates; a token "NT" in `rbp_cands` is not matched
}
\keyword{datasets}
