% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rbp_cands}
\alias{rbp_cands}
\title{a set of gene symbols used in the DESeq2 artifacts in ENCORE data resourceson AnVIL}
\format{
character()
}
\usage{
rbp_cands
}
\description{
a set of gene symbols used in the DESeq2 artifacts in ENCORE data resourceson AnVIL
}
\keyword{datasets}
